+++ modules/MetaFilters/iASamplingSettingsDlg.cpp
@@ -910,7 +910,7 @@ void iASamplingSettingsDlg::runClicked()
 				if (mainWnd)
 				{
 					auto child = mainWnd->activeMdiChild();
-					if (static_cast<int>(filter->requiredImages() + filter->requiredMeshes()) > child->dataSetMap().size())
+					if (static_cast<size_t>(filter->requiredImages() + filter->requiredMeshes()) > child->dataSetMap().size())
 					{
 						msg += QString("Filter requires more inputs (%1 image(s) + %2 mesh(es)) "
 							"than the number of datasets loaded in current child (%3)!\n")
